+++ gDRcore/R/add_annotation.R
@@ -16,7 +16,7 @@
 #'
 add_CellLine_annotation <- function(df_metadata,
                                     fill = NULL) {
-
+  
   # Assertions:
   stopifnot(inherits(df_metadata, "data.frame"))
   checkmate::assert_string(fill, null.ok = TRUE)
@@ -25,39 +25,40 @@ add_CellLine_annotation <- function(df_metadata,
   if (requireNamespace("gDRwrapper", quietly = TRUE)) {
     gDRwrapper::add_CellLine_annotation(df_metadata, fill)
   } else {
-  
+    
     DB_cellid_header <- "cell_line_identifier"
     DB_cell_annotate <- c("cell_line_name", "primary_tissue", "doubling_time",
                           "parental_identifier", "subtype")
     cellline <- gDRutils::get_env_identifiers("cellline")
     cellline_name <- gDRutils::get_env_identifiers("cellline_name")
     add_clid <- gDRutils::get_header("add_clid")
-  
+    
     # Read local cell_lines annotations
     annotationPackage <- ifelse(requireNamespace("gDRinternal", quietly = TRUE),
                                 "gDRinternal", "gDRtestData")
-    CLs_info <- data.table::fread(system.file("data", "cell_lines.csv", package = annotationPackage), fill = TRUE)
-    CLs_info <- CLs_info[, c(DB_cellid_header, DB_cell_annotate), with = FALSE]
-  
+    CLs_info <- read.csv(system.file("data", "cell_lines.csv", package = annotationPackage))
+    CLs_info <- CLs_info[, c(DB_cellid_header, DB_cell_annotate)]
+    CLs_info[, "doubling_time"] <- as.numeric(CLs_info[, "doubling_time"])
+    
     if (nrow(CLs_info) == 0) return(df_metadata)
-  
+    
     validatedCLs <- all(unique(df_metadata[[cellline]]) %in% CLs_info[[DB_cellid_header]])
     missingTblCellLines <- NULL
     if (!is.null(fill) && !validatedCLs) {
       missingTblCellLines <- data.table::data.table(parental_identifier = fill,
                                                     cell_line_name = fill,
                                                     cell_line_identifier = unlist(
-                                                      unique(df_metadata[, cellline, with = FALSE])),
+                                                      unique(df_metadata[, cellline])),
                                                     doubling_time = NA,
                                                     primary_tissue = fill,
                                                     subtype = fill)
-  
+      
     }
-  
+    
     if (any(!df_metadata$clid %in% CLs_info[[DB_cellid_header]]) && !is.null(missingTblCellLines)) {
       CLs_info <- rbind(CLs_info, missingTblCellLines)
     }
-  
+    
     colnames(CLs_info) <- unlist(c(cellline, add_clid, tail(DB_cell_annotate, 2)))
     CLIDs <- unique(df_metadata[[cellline]])
     bad_CL <- CLs_info[cellline %in% CLIDs][[cellline]]
@@ -65,15 +66,15 @@ add_CellLine_annotation <- function(df_metadata,
       futile.logger::flog.warn("Cell line ID %s not found in cell line database",
                                paste(CLIDs[bad_CL], collapse = " ; "))
       temp_CLIDs <- as.data.frame(matrix(ncol = length(c(cellline,
-                                                             add_clid)),
-                                  nrow = length(CLIDs[bad_CL])))
+                                                         add_clid)),
+                                         nrow = length(CLIDs[bad_CL])))
       colnames(temp_CLIDs) <- c(cellline,
                                 add_clid)
       
       temp_CLIDs[, cellline] <- temp_CLIDs[, cellline_name] <- CLIDs[bad_CL]
       CLs_info <- rbind(CLs_info, temp_CLIDs, fill = TRUE)
     }
-  
+    
     futile.logger::flog.info("Merge with Cell line info")
     nrows_df <- nrow(df_metadata)
     df_metadata <- base::merge(df_metadata, CLs_info, by = cellline, all.x = TRUE)
@@ -96,7 +97,7 @@ add_CellLine_annotation <- function(df_metadata,
 #'
 add_Drug_annotation <- function(df_metadata,
                                 fill = NULL) {
-
+  
   # Assertions:
   stopifnot(inherits(df_metadata, "data.frame"))
   checkmate::assert_string(fill, null.ok = TRUE)
@@ -104,21 +105,21 @@ add_Drug_annotation <- function(df_metadata,
   if (requireNamespace("gDRwrapper", quietly = TRUE)) {
     gDRwrapper::add_Drug_annotation(df_metadata, fill)
   } else {
-  
+    
     nrows_df <- nrow(df_metadata)
-  
+    
     DB_drug_identifier <- c("gnumber", "drug_name", "drug_moa")
     drug <- gDRutils::get_env_identifiers("drug")
     untreated_tag <- gDRutils::get_env_identifiers("untreated_tag")
     drugname <- gDRutils::get_env_identifiers("drugname")
     drug_moa <- gDRutils::get_env_identifiers("drug_moa")
     drugsTreated <- unique(df_metadata[[drug]])
-  
+    
     # Read local drugs annotations
     annotationPackage <- ifelse(requireNamespace("gDRinternal", quietly = TRUE),
                                 "gDRinternal", "gDRtestData")
-    Drug_info <- data.table::fread(system.file("data/drugs.csv", package = annotationPackage),
-                                   fill = TRUE, header = TRUE)
+    Drug_info <- read.csv(system.file("data/drugs.csv", package = annotationPackage),
+                          header = TRUE)
     Drug_info <- Drug_info[, c("gnumber", "drug_name", "drug_moa")]
     data.table::setnames(Drug_info, c("drug", "drug_name", "drug_moa"))
     drugsTreated <- drugsTreated[!drugsTreated %in% untreated_tag]
@@ -131,13 +132,13 @@ add_Drug_annotation <- function(df_metadata,
                                         drug_moa = fill
       )
     }
-  
-  
+    
+    
     if (nrow(Drug_info) == 0) {
       df_metadata[, drugname] <- df_metadata[, drug]
       return(df_metadata)
     }
-  
+    
     # -----------------------
     Drug_info$drug <- remove_drug_batch(Drug_info$drug)
     Drug_info <-
@@ -148,7 +149,7 @@ add_Drug_annotation <- function(df_metadata,
       ),
       Drug_info)
     Drug_info <- Drug_info[!duplicated(Drug_info[["drug"]]), ]
-    DrIDs <- unique(unlist(df_metadata[, grep(drug, colnames(df_metadata)), with = FALSE]))
+    DrIDs <- unique(unlist(df_metadata[, grep(drug, colnames(df_metadata))]))
     if (any(!remove_drug_batch(drugsTreated) %in% Drug_info$drug) && !is.null(missingTblDrugs)) {
       Drug_info <- rbind(Drug_info, data.table::setnames(
         missingTblDrugs[!(remove_drug_batch(missingTblDrugs$drug) %in% Drug_info$drug), ],
@@ -185,8 +186,8 @@ add_Drug_annotation <- function(df_metadata,
       Drug_info_ <- Drug_info
       colnames(Drug_info_)[colnames(Drug_info_)
                            %in% c(drugname, drug_moa)] <- paste0(c(drugname, drug_moa),
-                                                                     gsub(".*(_\\d)", "\\1",
-                                                                          colnames(df_metadata)[i]))
+                                                                 gsub(".*(_\\d)", "\\1",
+                                                                      colnames(df_metadata)[i]))
       futile.logger::flog.info("Merge with Drug_info for %s", colnames(df_metadata)[[i]])
       df_metadata[[paste0(colnames(df_metadata)[[i]], "_temp")]] <- remove_drug_batch(df_metadata[[i]])
       df_metadata <- base::merge(df_metadata, Drug_info_, by.x =
@@ -195,7 +196,7 @@ add_Drug_annotation <- function(df_metadata,
       df_metadata <- df_metadata[, .SD, .SDcols = !endsWith(names(df_metadata), "temp")]
     }
     stopifnot(nrows_df == nrow(df_metadata))
-  
+    
     return(df_metadata)
   }
 }