+++ gDRcore/R/assign_conditions.R
@@ -1,8 +1,8 @@
 #' .assign_conditions
 #'
-#' Assign the treated, untreated, and reference conditions to a data.frame
+#' Assign every row in a code{data.frame} to be a \code{"treated"}, \code{"untreated"}, or \code{"reference"} condition.
 #'
-#' @param df_ data.frame containing the \code{nested_identifiers}.
+#' @param df_ \code{data.frame} containing the \code{nested_identifiers}.
 #' @param nested_identifiers Character vector of column names in \code{df_}
 #' that will determine a condition type.
 #'
@@ -10,9 +10,9 @@
 #' whether the condition is a \code{"treated"} (drug treated), \code{"untreated"}, or \code{"reference"} entry.
 #' 
 #' @details Conditions are considered \code{"untreated"} when both \code{nested_identifiers}
-#' have a value of 0. When nested_identifiers are drugs, this represent entries where all drugs 
-#' are at 0 concentration. Conditions are considered \code{"reference"} when only one
-#' of the \code{nested_identifiers} is 0. In the context of drugs, this represents a single-agent 
+#' have a value of \code{0}. When nested_identifiers are drugs, this represent entries where all drugs 
+#' are at \code{0} concentration. Conditions are considered \code{"reference"} when only one
+#' of the \code{nested_identifiers} is \code{0}. In the context of drugs, this represents a single-agent 
 #' response.
 #'
 #' @export