+++ gDRcore/R/test.utils.R
@@ -14,7 +14,15 @@ test_synthetic_data <- function(original,
   if (inherits(data, "SummarizedExperiment")) {
     reprocessed <- data
   } else {
-    reprocessed <- gDRcore::runDrugResponseProcessingPipeline(data, override_untrt_controls = override_untrt_controls)
+    data_type <- ifelse(combo, "combo", "single-agent")
+    nested_identifiers <- if (combo) {
+      .get_default_combo_identifiers()
+    } else {
+      .get_default_single_agent_identifiers()
+      }
+    reprocessed <- gDRcore::runDrugResponseProcessingPipeline(data, override_untrt_controls = override_untrt_controls,
+                                                              data_type = data_type,
+                                                              nested_identifiers = nested_identifiers)
   }
   
   if (!is.null(override_untrt_controls)) {
@@ -88,7 +96,7 @@ test_synthetic_data <- function(original,
 test_synthetic_data2 <- function(original, long_df, dataName, asys = c("Normalized", "Averaged")) {
   nested_ids <- intersect(c("Concentration", "Concentration_2"), colnames(long_df))
   reprocessed <- create_and_normalize_SE(long_df, nested_identifiers = nested_ids, nested_confounders = "Barcode")
-  reprocessed <- average_SE(reprocessed, nested_identifiers = nested_ids)
+  reprocessed <- average_SE(reprocessed, series_identifiers = nested_ids)
   for (asy in asys) {
     o_df <- S4Vectors::DataFrame(gDRutils::convert_se_assay_to_dt(original, asy))
     n_df <- S4Vectors::DataFrame(gDRutils::convert_se_assay_to_dt(reprocessed, asy))